+++ R/ci.R
@@ -1,19 +1,20 @@
 #' Continuous integration setup and badges
 #'
 #' @description
-#' Sets up continuous integration (CI) services for an R package that is
-#' developed on GitHub or, perhaps, GitLab. CI services can run `R CMD check`
-#' automatically on various platforms, triggered by each push or pull request.
-#' These functions
+#'
+#' \lifecycle{soft-deprecated}
+#'
+#' Some of these functions are now soft-deprecated since the tidyverse team has
+#' started using [GitHub Actions (GHA)](https://github.com/features/actions) for
+#' continuous integration (CI). See [use_github_actions()] for help configuring
+#' GHA. GHA functionality in usethis is actively maintained and exercised, which
+#' is no longer true for Travis-CI or AppVeyor.
+#'
+#' Sets up third-party continuous integration (CI) services for an R package
+#' that is developed on GitHub or, perhaps, GitLab. These functions
 #' * Add service-specific configuration files and add them to `.Rbuildignore`.
 #' * Activate a service or give the user a detailed prompt.
 #' * Provide the markdown to insert a badge into README.
-#'
-#' Note that the tidyverse team now uses [GitHub
-#' Actions](https://github.com/features/actions) for continuous integration, so
-#' the usethis functions supporting GitHub Actions are now the most actively
-#' maintained. See [use_github_actions()] or [use_tidy_github_actions()] to
-#' learn more.
 #'
 #' @name ci
 #' @aliases NULL
@@ -31,7 +32,12 @@ NULL
 #' @rdname ci
 use_travis <- function(browse = rlang::is_interactive(),
                        ext = c("com", "org")) {
-  repo_spec <- get_repo_spec()
+  lifecycle::deprecate_soft(
+    when = "2.0.0",
+    what = "usethis::use_travis()",
+    with = "use_github_actions()"
+  )
+  repo_spec <- repo_spec()
   ext <- arg_match(ext)
   new <- use_template(
     "travis.yml",
@@ -54,7 +60,7 @@ use_travis <- function(browse = rlang::is_interactive(),
 #' @export
 #' @rdname ci
 use_travis_badge <- function(ext = c("com", "org"), repo_spec = NULL) {
-  repo_spec <- repo_spec %||% get_repo_spec()
+  repo_spec <- repo_spec %||% repo_spec()
   ext <- arg_match(ext)
   url <- glue("https://travis-ci.{ext}/{repo_spec}")
   img <- glue("{url}.svg?branch=master")
@@ -84,7 +90,12 @@ uses_travis <- function() {
 #' @export
 #' @rdname ci
 use_appveyor <- function(browse = rlang::is_interactive()) {
-  repo_spec <- get_repo_spec()
+  lifecycle::deprecate_soft(
+    when = "2.0.0",
+    what = "usethis::use_appveyor()",
+    with = "use_github_actions()"
+  )
+  repo_spec <- repo_spec()
   new <- use_template("appveyor.yml", ignore = TRUE)
   if (!new) {
     return(invisible(FALSE))
@@ -110,7 +121,7 @@ appveyor_activate <- function(browse = is_interactive()) {
 #' @export
 #' @rdname ci
 use_appveyor_badge <- function(repo_spec = NULL) {
-  repo_spec <- repo_spec %||% get_repo_spec()
+  repo_spec <- repo_spec %||% repo_spec()
   img <- glue(
     "https://ci.appveyor.com/api/projects/status/github/",
     "{repo_spec}?branch=master&svg=true"
@@ -154,7 +165,7 @@ use_gitlab_ci <- function() {
 #' @rdname ci
 use_circleci <- function(browse = rlang::is_interactive(),
                          image = "rocker/verse:latest") {
-  repo_spec <- get_repo_spec()
+  repo_spec <- repo_spec()
   use_directory(".circleci", ignore = TRUE)
   new <- use_template(
     "circleci-config.yml",
@@ -178,7 +189,7 @@ use_circleci <- function(browse = rlang::is_interactive(),
 #' @rdname ci
 #' @export
 use_circleci_badge <- function(repo_spec = NULL) {
-  repo_spec <- repo_spec %||% get_repo_spec()
+  repo_spec <- repo_spec %||% repo_spec()
   url <- glue("https://circleci.com/gh/{repo_spec}")
   img <- glue("{url}.svg?style=svg")
   use_badge("CircleCI build status", url, img)