+++ R/utils-github.R
@@ -14,7 +14,7 @@ github_remote_protocol <- function(name = "origin") {
   protocol
 }
 
-## repo_spec --> owner, repo
+# repo_spec --> owner, repo
 parse_repo_spec <- function(repo_spec) {
   repo_split <- strsplit(repo_spec, "/")[[1]]
   if (length(repo_split) != 2) {
@@ -26,7 +26,13 @@ parse_repo_spec <- function(repo_spec) {
 spec_owner <- function(repo_spec) parse_repo_spec(repo_spec)$owner
 spec_repo <- function(repo_spec) parse_repo_spec(repo_spec)$repo
 
-## named vector or list of GitHub URLs --> data frame of URL parts
+# owner, repo --> repo_spec
+make_spec <- function(owner = NA, repo = NA) {
+  no_spec <- is.na(owner) | is.na(repo)
+  ifelse(no_spec, NA_character_, glue("{owner}/{repo}"))
+}
+
+# named vector or list of GitHub URLs --> data frame of URL parts
 parse_github_remotes <- function(x) {
   # https://github.com/r-lib/usethis             --> https, rlib, usethis
   # https://github.com/r-lib/usethis.git         --> https, rlib, usethis
@@ -77,8 +83,10 @@ github_remote_from_description <- function(desc) {
 #' @param github_get Whether to use the `/repos/:owner/:repo` endpoint of the
 #'   GitHub API to get more information about each remote repository, such as
 #'   whether it's a fork and whether the user can push. `NA` (the default) means
-#'   to add this info if possible. An explicit `TRUE` or `FALSE` is taken
-#'   literally. *Not fully implemented yet.*
+#'   to add this info if: (1) we are online and (2) we have a PAT. An explicit
+#'   `TRUE` or `FALSE` is taken literally. Technically, for public repos, we can
+#'   get most of this data from the API without a PAT, but until I have reason
+#'   to plan for that scenario, I choose not to bother.
 #' @inheritParams use_github
 #' @keywords internal
 #' @noRd
@@ -86,9 +94,9 @@ github_remotes <- function(these = c("origin", "upstream"),
                            github_get = NA,
                            auth_token = github_token(),
                            host = "https://api.github.com") {
-  # TODO: Fully implement the behaviour promised for `github_get`:
-  # * no internet access
-  # * no GITHUB_PAT
+  if (is.na(github_get)) {
+    github_get <- is_online("github.com") && have_github_token(auth_token)
+  }
   if (isTRUE(github_get)) {
     check_github_token(auth_token)
   }
@@ -97,10 +105,13 @@ github_remotes <- function(these = c("origin", "upstream"),
     url = NA_character_,
     repo_owner = NA_character_,
     repo_name = NA_character_,
+    repo_spec = NA_character_,
+    github_get = NA,
     is_fork = NA,
     can_push = NA,
     parent_repo_owner = NA_character_,
     parent_repo_name = NA_character_,
+    parent_repo_spec = NA_character_,
     can_push_to_parent = NA,
     stringsAsFactors = FALSE
   )
@@ -113,6 +124,7 @@ github_remotes <- function(these = c("origin", "upstream"),
     return(grl[0, ])
   }
   grl <- grl[rep_len(1, nrow(x)), ]
+  grl$github_get <- github_get
 
   grl$remote <- x$name
   grl$url <- x$url
@@ -120,8 +132,9 @@ github_remotes <- function(these = c("origin", "upstream"),
   parsed <- parse_github_remotes(grl$url)
   grl$repo_owner <- parsed$repo_owner
   grl$repo_name <- parsed$repo_name
+  grl$repo_spec <-  make_spec(grl$repo_owner, grl$repo_name)
 
-  if (identical(github_get, FALSE)) {
+  if (!github_get) {
     return(grl)
   }
 
@@ -139,13 +152,14 @@ github_remotes <- function(these = c("origin", "upstream"),
 
   repo_info <- purrr::map2(grl$repo_owner, grl$repo_name, get_gh_repo)
   grl$is_fork <- map_lgl(repo_info, "fork")
-  # TODO: permission won't be here if we don't have a PAT
-  # deal with that
-  grl$can_push <- map_lgl(repo_info, c("permissions", "push"))
+  # `permissions` is an example of data that is not present if the request
+  # did not include a PAT
+  grl$can_push <- map_lgl(repo_info, c("permissions", "push"), .default = NA)
   grl$parent_repo_owner <-
     map_chr(repo_info, c("parent", "owner", "login"), .default = NA)
   grl$parent_repo_name <-
     map_chr(repo_info, c("parent", "name"), .default = NA)
+  grl$parent_repo_spec <-  make_spec(grl$parent_repo_owner, grl$parent_repo_name)
 
   parent_info <- purrr::map2(
     grl$parent_repo_owner, grl$parent_repo_name,
@@ -161,99 +175,123 @@ github_remotes <- function(these = c("origin", "upstream"),
 #'
 #' @description
 #' Classify the active project's GitHub remote situation, so diagnostic and
-#' other downstream functions can decide whether to scold/fix/proceed/abort.
+#' other downstream functions can decide whether to proceed / abort / complain &
+#' offer to fix.
 #' We only consider the remotes where:
 #' * Name is `origin` or `upstream` and URL contains the string `"github"`
 #'
 #' We assume the project is a Git repo, so use this behind a guard like
 #' `check_uses_git()`.
 #'
-#' Some setups are *supported*, meaning usethis can do its job correctly, and
-#' others are *unsupported*, meaning usethis can't fulfill the user's request.
-#' Some *supported* setups are still considered suboptimal. Here's how I think
-#' this will work for any specific usethis function:
-#' * `stop_bad_github_config()` for unsupported setups
-#' * Use `ui_github_config_wat()` to inform the user about
-#'   suboptimal-but-supported setups and give them a chance to back out. The
-#'   setups that trigger this may vary from function to function.
-#' * Proceed quietly for setups we consider ideal.
+#' We have to call the GitHub API to fully characterize the GitHub remote
+#' situation. That's the only way to learn if the user can push to a remote,
+#' whether a remote is a fork, and which repo is the parent of a fork.
+#' `github_get` controls whether we make these API calls.
+#'
+#' Some functions can get by with the information that's available locally, i.e.
+#' we can use simple logic to decide whether to target `origin` or `upstream` or
+#' present the user with a choice. We can set `github_get = FALSE` in this case.
+#' Other functions, like the `pr_*()` functions, are more demanding and we'll
+#' always determine the config with `github_get = TRUE`.
+#'
+#' Here's how a usethis function can use the GitHub remote configuration:
+#' * `cfg <- classify_github_config(...)`
+#' * Inspect `cfg$type` and call `stop_bad_github_remote_config()` if the
+#'   function can't work with that config.
+#' * If the config is suboptimal-but-supported, use
+#'   `ui_github_remote_config_wat()` to educate the user and give them a chance
+#'   to back out.
+#' * Proceed quietly if the config is OK.
+#'
+#' Fields in an instance of `github_remote_config`:
+#' * `type`: explained below
+#' * `pr_ready`: Logical. Do the `pr_*()` functions support it?
+#' * `desc`: A description used in messages and menus.
+#' * `origin`: Information about the `origin` GitHub remote.
+#' * `upstream`: Information about the `origin` GitHub remote.
 #'
 #' Possible GitHub remote configurations, the common cases:
-#' * no_github: No `origin`, no `upstream`. Unsupported.
+#' * no_github: No `origin`, no `upstream`.
 #' * ours: `origin` exists, is not a fork, and we can push to it. Owner of
 #'   `origin` could be current user, another user, or an org. No `upstream`.
-#' * theirs: `origin` exists, is not a fork, and we can NOT push to it. No
-#'   `upstream`.
+#'   - Less common variant: `upstream` exists, `origin` does not, and we can
+#'     push to `upstream`. The fork-ness of `upstream` is not consulted.
 #' * fork: `origin` exists and we can push to it. `origin` is a fork of the repo
 #'   configured as `upstream`. We may or may not be able to push to `upstream`.
-#' * fork_no_upstream: `origin` exists and we can push to it. `origin` is a
-#'   fork, but that parent repo has not (yet) been configured as the
-#'   `upstream` remote.
+#' * theirs: Exactly one of `origin` and `upstream` exist and we can't push to
+#'   it. The fork-ness of this remote repo is not consulted.
 #'
 #' Possible GitHub remote configurations, the peculiar ones:
-#' * upstream_only: `upstream` exists but `origin` does not.
-#' * fork_origin_read_only: `origin` exists, it's a fork, but we can't push to
-#'   `origin`. `upstream` may or may not be configured.
-#' * fork_upstream_is_not_origin_parent: `origin` exists, it's a fork of
-#'   something, `upstream` exists, but `origin` is not a fork of `upstream`.
+#' * fork_upstream_is_not_origin_parent: `origin` exists, it's a fork, but its
+#'   parent repo is not configured as `upstream`. Either there's no `upstream`
+#'   or `upstream` exists but it's not the parent of `origin`.
+#' * fork_cannot_push_origin: `origin` is a fork and its parent is configured
+#'   as `upstream`. But we can't push to `origin`.
 #' * upstream_but_origin_is_not_fork: `origin` and `upstream` both exist, but
 #'   `origin` is not a fork of anything and, specifically, it's not a fork of
 #'   `upstream`.
 #'
-#' @inheritParams use_github
+#'  Remote configuration "guesses" we apply when `github_get = FALSE`:
+#'  * maybe_ours_or_theirs: Exactly one of `origin` and `upstream` exists.
+#'  * maybe_fork: Both `origin` and `upstream` exist.
+#'
+#' @inheritParams github_remotes
 #' @keywords internal
 #' @noRd
-classify_github_setup <- function(auth_token = github_token(),
-                                  host = "https://api.github.com") {
-  cfg <- new_github_config()
-  grl <- github_remotes(auth_token = auth_token, host = host, github_get = TRUE)
+github_remote_config <- function(github_get = NA,
+                                 auth_token = github_token(),
+                                 host = "https://api.github.com") {
+  cfg <- new_github_remote_config()
+  grl <- github_remotes(
+    github_get = github_get,
+    auth_token = auth_token,
+    host = host
+  )
 
   if (nrow(grl) == 0) {
     return(cfg_no_github(cfg))
   }
 
+  github_get <- grl$github_get[1] # assuming it's same for origin and upstream
   cfg$origin$is_configured   <- "origin"   %in% grl$remote
   cfg$upstream$is_configured <- "upstream" %in% grl$remote
-
-  make_spec <- function(owner = NA, repo = NA) {
-    if (is.na(owner) || is.na(repo)) {
-      NA_character_
-    } else {
-      glue("{owner}/{repo}")
-    }
-  }
 
   if (cfg$origin$is_configured) {
     cols <- intersect(names(grl), names(cfg$origin))
     origin <- grl[grl$remote == "origin", cols]
     cfg$origin <- utils::modifyList(cfg$origin, origin)
-    cfg$origin$repo_spec <- with(cfg$origin, make_spec(repo_owner, repo_name))
-    cfg$origin$parent_repo_spec <-
-      with(cfg$origin, make_spec(parent_repo_owner, parent_repo_name))
   }
 
   if (cfg$upstream$is_configured) {
     cols <- intersect(names(grl), names(cfg$upstream))
     upstream <- grl[grl$remote == "upstream", cols]
     cfg$upstream <- utils::modifyList(cfg$upstream, upstream)
-    cfg$upstream$repo_spec <- with(cfg$upstream, make_spec(repo_owner, repo_name))
   }
+  # cfg is as complete as it can be if `github_get` is `FALSE`
 
-  if (cfg$upstream$is_configured && cfg$origin$is_configured) {
-    cfg$origin$parent_is_upstream <- identical(
-      cfg$origin$parent_repo_spec,
-      cfg$upstream$repo_spec
-    )
+  single_remote <- xor(cfg$origin$is_configured, cfg$upstream$is_configured)
+
+  if (!isTRUE(github_get)) {
+    if (single_remote) {
+      return(cfg_maybe_ours_or_theirs(cfg))
+    } else {
+      return(cfg_maybe_fork(cfg))
+    }
   }
-  # cfg is now fully populated, up to type / unsupported / desc
+  # `github_get` must be `TRUE`
+
+  if (!single_remote) {
+    cfg$origin$parent_is_upstream <-
+      identical(cfg$origin$parent_repo_spec, cfg$upstream$repo_spec)
+  }
 
   # origin only
-  if (cfg$origin$is_configured && !cfg$upstream$is_configured) {
+  if (single_remote && cfg$origin$is_configured) {
     if (cfg$origin$is_fork) {
       if (cfg$origin$can_push) {
-        return(cfg_fork_no_upstream(cfg))
+        return(cfg_fork_upstream_is_not_origin_parent(cfg))
       } else {
-        return(cfg_fork_origin_read_only(cfg))
+        return(cfg_theirs(cfg))
       }
     } else {
       if (cfg$origin$can_push) {
@@ -265,31 +303,35 @@ classify_github_setup <- function(auth_token = github_token(),
   }
 
   # upstream only
-  if (!cfg$origin$is_configured && cfg$upstream$is_configured) {
-    return(cfg_upstream_only(cfg))
+  if (single_remote && cfg$upstream$is_configured) {
+    if (cfg$upstream$can_push) {
+      return(cfg_ours(cfg))
+    } else {
+      return(cfg_theirs(cfg))
+    }
   }
 
   # origin and upstream
   if (cfg$origin$is_fork) {
-    if (cfg$origin$can_push) {
-      if (cfg$origin$parent_is_upstream) {
+    if (cfg$origin$parent_is_upstream) {
+      if (cfg$origin$can_push) {
         return(cfg_fork(cfg))
       } else {
-        return(cfg_fork_upstream_is_not_origin_parent(cfg))
+        return(cfg_fork_cannot_push_origin(cfg))
       }
     } else {
-      return(cfg_fork_origin_read_only(cfg))
+      return(cfg_fork_upstream_is_not_origin_parent(cfg))
     }
   } else {
     return(cfg_upstream_but_origin_is_not_fork(cfg))
   }
 }
 
-new_github_config <- function() {
+new_github_remote_config <- function() {
   structure(
     list(
       type = NA_character_,
-      unsupported = TRUE,
+      pr_ready = FALSE,
       desc = "Unexpected remote configuration.",
       origin = list(
         name = "origin",
@@ -316,120 +358,105 @@ new_github_config <- function() {
         repo_spec = NA_character_
       )
     ),
-    class = "github_config"
+    class = "github_remote_config"
   )
 }
 
 #' Get a repo spec
 #'
-#' Returns an OWNER/REPO repo spec for, e.g., incorporating into a URL.
+#' @description
+#' Returns a repo spec of the form OWNER/REPO for, e.g., incorporating into a URL.
 #' Examples:
 #' * Badge URLs
 #' * URLs where you can activate a CI service
-#' * URLs for URL and BugReports in DESCRIPTION
+#' * URLs for DESCRIPTION fields such as URL and BugReports
 #'
-#' Use this when you suspect the user is doing something in order to make a pull
-#' request or if it's not clear if they want to work against their fork or the
-#' fork parent. `get_repo_spec()` will challenge "theirs" and "fork_no_upstream"
-#' and ask if user wants to back out and fix the remote configuration. In the
-#' case of "fork" and "fork_no_upstream", user gets an interactive menu to
-#' choose between `origin` and its parent repo (which may or may not be
-#' configured as `upstream`).
+#' If `cfg` is not provided, `repo_spec()` calls
+#' `github_remote_config(github_get = FALSE)` and works only with locally
+#' available information about GitHub remotes. To work with full GitHub remote
+#' configuration, call `github_remote_config(github_get = TRUE)` yourself and
+#' pass the resulting `cfg` in. `repo_spec()` will challenge certain configs,
+#' e.g., "fork_upstream_is_not_origin_parent", and ask if user wants to back out
+#' and fix the remote configuration.
+#'
+#' In some configurations, if `ask = TRUE` and we're in an interactive session,
+#' user gets a choice between `origin` and (if either exists or is known) its
+#' parent repo and `upstream`.
 #'
 #' @inheritParams use_github
 #' @keywords internal
 #' @noRd
-get_repo_spec <- function(auth_token = github_token(),
-                          host = "https://api.github.com") {
-  cfg <- classify_github_setup(auth_token = auth_token, host = host)
-  if (cfg$unsupported) {
-    stop_bad_github_config(cfg)
+repo_spec <- function(cfg = NULL,
+                      role = c("source", "primary"),
+                      ask = is_interactive(),
+                      auth_token = github_token(),
+                      host = "https://api.github.com") {
+  cfg <- cfg %||%
+    github_remote_config(github_get = FALSE, auth_token = auth_token, host = host)
+  stopifnot(inherits(cfg, "github_remote_config"))
+  role <- match.arg(role)
+
+  if (cfg$type == "no_github") {
+    stop_bad_github_remote_config(cfg)
   }
-  if (cfg$type == "ours") {
-    return(cfg$origin$repo_spec)
-  }
-  if (cfg$type == "theirs") {
-    if (ui_github_config_wat(cfg)) {
+
+  weird_configs <- c(
+    "fork_upstream_is_not_origin_parent",
+    "fork_cannot_push_origin",
+    "upstream_but_origin_is_not_fork"
+  )
+  if (is_interactive() && cfg$type %in% weird_configs) {
+    if (ui_github_remote_config_wat(cfg)) {
       ui_stop("Exiting due to unfavorable GitHub config")
-    } else {
+    }
+  }
+
+  # upstream only
+  if (cfg$upstream$is_configured && !cfg$origin$is_configured) {
+    return(cfg$upstream$repo_spec)
+  }
+
+  # origin only
+  if (cfg$origin$is_configured && !cfg$upstream$is_configured) {
+    if (is.na(cfg$origin$parent_repo_spec)) {
       return(cfg$origin$repo_spec)
     }
   }
-  if (cfg$type == "fork_no_upstream") {
-    if (ui_github_config_wat(cfg)) {
-      ui_stop("Exiting due to unfavorable GitHub config")
-    }
-  }
-  if (!(cfg$type %in% c("fork", "fork_no_upstream"))) {
-    ui_stop("Internal error. Unexpected GitHub config type: {cfg$type}")
-  }
-  if (!is_interactive()) {
-    ui_info("
-      Working with a fork, non-interactively. \\
-      Targetting the fork parent = {ui_value(cfg$origin$parent_repo_spec)}
-      ")
-    cfg$origin$parent_repo_spec
-  } else {
-    choices <- c(
-      glue("{cfg$origin$parent_repo_spec} = parent of your fork"),
-      glue("{cfg$origin$repo_spec} = your fork")
-    )
-    title <- glue("
-      Working with a fork.
-      Which repo should we target?
-      ")
-    choice <- utils::menu(choices, graphics = FALSE, title = title)
-    return(with(cfg$origin, c(parent_repo_spec, repo_spec)[choice]))
-  }
-}
+  # scenarios left (X means "not NA", - means "is NA"):
+  # origin origin_parent upstream
+  #   X          X          -
+  #   X          -          X
+  #   X          X    ==    X
+  #   X          X    !=    X
 
-# Like get_repo_spec(), but do much less. No API calls.
-# No full-fledged evaluation of the GitHub remotes. Take them at face value.
-get_repo_spec_lite <- function() {
-  grl <- github_remotes(github_get = FALSE)
-  if (nrow(grl) < 1) {
-    return()
+  if (!ask || !is_interactive()) {
+    return(switch(
+      role,
+      source  = cfg$origin$parent_repo_spec %|% cfg$upstream$repo_spec,
+      primary = cfg$origin$repo_spec
+    ))
   }
-  grl$repo_spec <- glue_data(grl, "{repo_owner}/{repo_name}")
-  if (!is_interactive() && nrow(grl) > 1) {
-    ui_info("
-      Both {ui_value('origin')} and {ui_value('upstream')} are GitHub remotes.
-      Using spec for {ui_value('origin')}.")
-    return(grl$repo_spec[grl$remote == "origin"])
-  }
-  if (nrow(grl) == 1) {
-    return(grl$repo_spec)
-  }
-  choice <- utils::menu(
-    choices = glue_data(grl, "{repo_spec} ({remote})"),
-    graphics = FALSE,
-    title = "Which repo should we target?"
+
+  spec <- list(
+    origin        = cfg$origin$repo_spec,
+    origin_parent = cfg$origin$parent_repo_spec,
+    upstream      = cfg$upstream$repo_spec
   )
-  grl$repo_spec[choice]
-}
+  formatted <- c(
+    origin        = glue("{spec$origin} = {ui_value('origin')}"),
+    origin_parent = glue("{spec$origin_parent} = parent of {ui_value('origin')}"),
+    upstream      = glue("{spec$upstream} = {ui_value('upstream')}")
+  )
 
-#' Get the spec of the primary repo
-#'
-#' This is a simplified variant of `get_repo_spec()`. Use it when you're quite
-#' sure that you want the primary repo (so fork parent, in the case of a fork),
-#' with no nudges or interactive choice.
-#'
-#' @keywords internal
-#' @noRd
-get_primary_spec <- function() {
-  cfg <- classify_github_setup()
-  if (cfg$unsupported) {
-    stop_bad_github_config(cfg)
+  spec <- spec[!is.na(spec)]
+  if (length(spec) == 3 &&
+      identical(spec[["origin_parent"]], spec[["upstream"]])) {
+    spec <- spec[c("origin", "upstream")]
   }
-  if (cfg$type %in% c("ours", "theirs")) {
-    cfg$origin$repo_spec
-  } else if (cfg$type %in% c("fork", "fork_no_upstream")) {
-    out <- cfg$origin$parent_repo_spec
-    ui_info("Targetting the fork parent = {ui_value(out)}")
-    out
-  } else {
-    ui_stop("Internal error. Unexpected GitHub config type: {cfg$type}")
-  }
+  choices <- formatted[names(spec)]
+  title <- glue("Which repo should we target?")
+  choice <- utils::menu(choices, graphics = FALSE, title = title)
+  spec[[choice]]
 }
 
 # formatting github remote configurations for humans ---------------------------
@@ -442,9 +469,10 @@ format_remote <- function(remote) {
     }
   }
   push_clause <- function(remote) {
-    if (remote$is_configured) {
-      if (remote$can_push) " (can push)" else " (can not push)"
+    if (!remote$is_configured || is.na(remote$can_push)) {
+      return()
     }
+    if (remote$can_push) " (can push)" else " (can not push)"
   }
   out <- c(
     glue("{remote$name} = {effective_spec(remote)}"),
@@ -459,7 +487,7 @@ format_remote <- function(remote) {
 format_fields <- function(cfg) {
   list(
     type = glue("type = {ui_value(cfg$type)}"),
-    unsupported = glue("unsupported = {ui_value(cfg$unsupported)}"),
+    pr_ready = glue("Config supports a pull request = {ui_value(cfg$pr_ready)}"),
     origin = format_remote(cfg$origin),
     upstream = format_remote(cfg$upstream),
     desc = if (is.na(cfg$desc)) {
@@ -471,23 +499,23 @@ format_fields <- function(cfg) {
 }
 
 #' @export
-format.github_config <- function(x, ...) {
+format.github_remote_config <- function(x, ...) {
   glue::as_glue(format_fields(x))
 }
 
 #' @export
-print.github_config <- function(x, ...) {
+print.github_remote_config <- function(x, ...) {
   cat(format(x, ...), sep = "\n")
   invisible(x)
 }
 
 # refines output of format_fields() to create input better suited to
-# ui_github_config_wat() and stop_bad_github_config()
-github_config_wat <- function(cfg, context = c("menu", "abort")) {
+# ui_github_remote_config_wat() and stop_bad_github_remote_config()
+github_remote_config_wat <- function(cfg, context = c("menu", "abort")) {
   context <- match.arg(context)
   adjective <- switch(context, menu = "Unexpected", abort = "Unsupported")
   out <- format_fields(cfg)
-  out$unsupported <- NULL
+  out$pr_ready <- NULL
   out$type <- glue("{adjective} GitHub remote configuration: {ui_value(cfg$type)}")
   out$desc <- if (is.na(cfg$desc)) NULL else cfg$desc
   out
@@ -495,29 +523,28 @@ github_config_wat <- function(cfg, context = c("menu", "abort")) {
 
 # returns TRUE if user selects "no" --> exit the calling function
 # return FALSE if user select "yes" --> keep going, they've been warned
-ui_github_config_wat <- function(cfg) {
+ui_github_remote_config_wat <- function(cfg) {
   ui_nope(
-    github_config_wat(cfg, context = "menu"),
+    github_remote_config_wat(cfg, context = "menu"),
     yes = "Yes, I want to proceed. I know what I'm doing.",
     no = "No, I want to stop and straighten out my GitHub remotes first.",
     shuffle = FALSE
   )
 }
 
-stop_bad_github_config <- function(cfg) {
+stop_bad_github_remote_config <- function(cfg) {
   abort(
-    message = unname(github_config_wat(cfg, context = "abort")),
-    class = c("usethis_error_bad_github_config", "usethis_error"),
+    message = unname(github_remote_config_wat(cfg, context = "abort")),
+    class = c("usethis_error_bad_github_remote_config", "usethis_error"),
     cfg = cfg
   )
 }
 
 stop_unsupported_pr_config <- function(cfg) {
-  msg <- github_config_wat(cfg)
+  msg <- github_remote_config_wat(cfg)
   msg$type <- glue("
     Pull request functions can't work with GitHub remote configuration: \\
-    {ui_value(cfg$type)}
-    ")
+    {ui_value(cfg$type)}")
   abort(
     message = unname(msg),
     class = c("usethis_error_invalid_pr_config", "usethis_error"),
@@ -525,13 +552,13 @@ stop_unsupported_pr_config <- function(cfg) {
   )
 }
 
-# common configurations --------------------------------------------------------
+# github remote configurations -------------------------------------------------
 cfg_no_github <- function(cfg) {
   utils::modifyList(
     cfg,
     list(
       type = "no_github",
-      unsupported = TRUE,
+      pr_ready = FALSE,
       desc = glue("
         Neither {ui_value('origin')} nor {ui_value('upstream')} is a GitHub \\
         repo.")
@@ -540,22 +567,47 @@ cfg_no_github <- function(cfg) {
 }
 
 cfg_ours <- function(cfg) {
-  utils::modifyList(cfg, list(type = "ours", unsupported = FALSE, desc = NA))
+  utils::modifyList(
+    cfg,
+    list(
+      type = "ours",
+      pr_ready = TRUE,
+      desc = NA)
+  )
 }
 
 cfg_theirs <- function(cfg) {
+  configured <- if (cfg$origin$is_configured) "origin" else "upstream"
   utils::modifyList(
     cfg,
     list(
       type = "theirs",
-      unsupported = FALSE,
+      pr_ready = FALSE,
       desc = glue("
-        The only configured GitHub remote is {ui_value('origin')}, which you \\
-        cannot push to.
-        If your goal is to make a pull request, you must \\
-        fork-and-clone.
-        {ui_code('usethis::create_from_github()')} can do this.
-        ")
+        The only configured GitHub remote is {ui_value(configured)}, which
+        you cannot push to.
+        If your goal is to make a pull request, you must fork-and-clone.
+        {ui_code('usethis::create_from_github()')} can do this.")
+    )
+  )
+}
+
+cfg_maybe_ours_or_theirs <- function(cfg) {
+  if (cfg$origin$is_configured) {
+    configured <- "origin"
+    not_configured <- "upstream"
+  } else {
+    configured <- "upstream"
+    not_configured <- "origin"
+  }
+  utils::modifyList(
+    cfg,
+    list(
+      type = "maybe_ours_or_theirs",
+      pr_ready = NA,
+      desc = glue("
+        {ui_value(configured)} is a GitHub repo and {ui_value(not_configured)} \\
+        is either not configured or is not a GitHub repo.")
     )
   )
 }
@@ -565,64 +617,36 @@ cfg_fork <- function(cfg) {
     cfg,
     list(
       type = "fork",
-      unsupported = FALSE,
+      pr_ready = TRUE,
       # TODO: say whether user can push to parent / upstream?
       desc = glue("
         {ui_value('origin')} is a fork of {ui_value(cfg$upstream$repo_spec)}, \\
-        which is configured as the {ui_value('upstream')} remote.
-        ")
-    ))
-}
-
-cfg_fork_no_upstream <- function(cfg) {
-  utils::modifyList(
-    cfg,
-    list(
-      type = "fork_no_upstream",
-      unsupported = FALSE, # I'm ambivalent about this. We CAN do some things in
-      # this situation, but it's dysfunctional long-term,
-      # because we won't be able to pull from upstream.
-      desc = glue("
-        {ui_value('origin')} is a fork, but its parent is not configured \\
-        as {ui_value('upstream')}.
-        You can make a pull request.
-        However, going forward, you can't pull changes from the main repo.
-        Use {ui_code('usethis::use_git_remote()')} to add the parent repo as \\
-        the {ui_value('upstream')} remote.
-        ")
+        which is configured as the {ui_value('upstream')} remote.")
     )
   )
 }
 
-# peculiar configurations ------------------------------------------------------
-cfg_fork_origin_read_only <- function(cfg) {
+cfg_maybe_fork <- function(cfg) {
   utils::modifyList(
     cfg,
     list(
-      type = "fork_origin_read_only",
-      unsupported = TRUE, # I'm ambivalent about this, I suppose we could view
-                          # this as an unusual version of "theirs"? We do
-                          # actually know the parent repo, so it's possible to
-                          # do some local work. But then we can't push it.
+      type = "maybe_fork",
+      pr_ready = NA,
+      desc = glue("
+        Both {ui_value('origin')} and {ui_value('upstream')} appear to be \\
+        GitHub repos.")
+    )
+  )
+}
+
+cfg_fork_cannot_push_origin <- function(cfg) {
+  utils::modifyList(
+    cfg,
+    list(
+      type = "fork_cannot_push_origin",
+      pr_ready = FALSE,
       desc = glue("
         The {ui_value('origin')} remote is a fork, but you can't push to it.")
-    )
-  )
-}
-
-cfg_upstream_only <- function(cfg) {
-  utils::modifyList(
-    cfg,
-    list(
-      type = "upstream_only",
-      unsupported = TRUE, # I'm ambivalent about this. If we can push, this could
-                          # be regarded as yet another look for "ours". If we
-                          # can't push, this could be regarded as variant of
-                          # "theirs".
-      desc = glue("
-        The only GitHub remote is {ui_value('upstream')}.
-        usethis expects {ui_value('origin')} or both {ui_value('origin')} and \\
-        {ui_value('upstream')} to be configured.")
     )
   )
 }
@@ -632,10 +656,10 @@ cfg_fork_upstream_is_not_origin_parent <- function(cfg) {
     cfg,
     list(
       type = "fork_upstream_is_not_origin_parent",
-      unsupported = TRUE,
+      pr_ready = FALSE,
       desc = glue("
-        The {ui_value('origin')} GitHub remote is a fork, but it's not a fork \\
-        of the repo configured as the {ui_value('upstream')} remote.")
+        The {ui_value('origin')} GitHub remote is a fork, but its parent is \\
+        not configured as the {ui_value('upstream')} remote.")
     )
   )
 }
@@ -645,7 +669,7 @@ cfg_upstream_but_origin_is_not_fork <- function(cfg) {
     cfg,
     list(
       type = "upstream_but_origin_is_not_fork",
-      unsupported = TRUE,
+      pr_ready = FALSE,
       desc = glue("
         Both {ui_value('origin')} and {ui_value('upstream')} are GitHub \\
         remotes, but {ui_value('origin')} is not a fork and, in particular, \\
@@ -653,4 +677,3 @@ cfg_upstream_but_origin_is_not_fork <- function(cfg) {
     )
   )
 }
-